+++ docs/supported-libraries.md
@@ -138,7 +138,7 @@ These are the supported libraries and frameworks:
 | [Spring Integration](https://spring.io/projects/spring-integration)                                                                         | 4.1+ (not including 6.0+ yet)      | [opentelemetry-spring-integration-4.1](../instrumentation/spring/spring-integration-4.1/library)                                                                                                                                                                                                                                                                                        | [Messaging Spans]                                                                                                                       |
 | [Spring JMS](https://docs.spring.io/spring-framework/docs/current/reference/html/integration.html#jms)                                      | 2.0+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [Messaging Spans]                                                                                                                       |
 | [Spring Kafka](https://spring.io/projects/spring-kafka)                                                                                     | 2.7+                               | [opentelemetry-spring-kafka-2.7](../instrumentation/spring/spring-kafka-2.7/library)                                                                                                                                                                                                                                                                                                    | [Messaging Spans]                                                                                                                       |
-| [Spring Pulsar](https://spring.io/projects/spring-pulsar)                                                                                   | 1.2+                               |                                                                                                                                                                                                                                                                                                                                                                                         | [Messaging Spans]                                                                                                                       |
+| [Spring Pulsar](https://spring.io/projects/spring-pulsar)                                                                                   | 1.0+                               |                                                                                                                                                                                                                                                                                                                                                                                         | [Messaging Spans]                                                                                                                       |
 | [Spring RabbitMQ](https://spring.io/projects/spring-amqp)                                                                                   | 1.0+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [Messaging Spans]                                                                                                                       |
 | [Spring RestTemplate](https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/web/client/package-summary.html) | 3.1+                               | [opentelemetry-spring-web-3.1](../instrumentation/spring/spring-web/spring-web-3.1/library)                                                                                                                                                                                                                                                                                             | [HTTP Client Spans], [HTTP Client Metrics]                                                                                              |
 | [Spring RMI](https://docs.spring.io/spring-framework/docs/4.0.x/javadoc-api/org/springframework/remoting/rmi/package-summary.html)          | 4.0+                               | N/A                                                                                                                                                                                                                                                                                                                                                                                     | [RPC Client Spans], [RPC Server Spans]                                                                                                  |
