--- instrumentation/spring/spring-pulsar-1.2/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/pulsar/v1_2/DefaultPulsarMessageListenerContainerInstrumentation.java
@@ -1,66 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.spring.pulsar.v1_2;
-
-import static io.opentelemetry.javaagent.instrumentation.spring.pulsar.v1_2.SpringPulsarSingletons.instrumenter;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import io.opentelemetry.javaagent.instrumentation.pulsar.v2_8.VirtualFieldStore;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.pulsar.client.api.Message;
-
-public class DefaultPulsarMessageListenerContainerInstrumentation implements TypeInstrumentation {
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return named(
-        "org.springframework.pulsar.listener.DefaultPulsarMessageListenerContainer$Listener");
-  }
-
-  @Override
-  public void transform(TypeTransformer transformer) {
-    transformer.applyAdviceToMethod(
-        named("dispatchMessageToListener")
-            .and(takesArguments(3))
-            .and(takesArgument(0, named("org.apache.pulsar.client.api.Message"))),
-        getClass().getName() + "$DispatchMessageToListenerAdvice");
-  }
-
-  @SuppressWarnings("unused")
-  public static class DispatchMessageToListenerAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void onEnter(
-        @Advice.Argument(0) Message<?> message,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = VirtualFieldStore.extract(message);
-      if (instrumenter().shouldStart(parentContext, message)) {
-        context = instrumenter().start(parentContext, message);
-        scope = context.makeCurrent();
-      }
-    }
-
-    @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class)
-    public static void onExit(
-        @Advice.Argument(0) Message<?> message,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope,
-        @Advice.Thrown Throwable throwable) {
-      if (scope == null) {
-        return;
-      }
-      scope.close();
-      instrumenter().end(context, message, null, throwable);
-    }
-  }
-}
