--- instrumentation/spring/spring-pulsar-1.2/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/pulsar/v1_2/SpringPulsarInstrumentationModule.java
@@ -1,25 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.spring.pulsar.v1_2;
-
-import static java.util.Collections.singletonList;
-
-import com.google.auto.service.AutoService;
-import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import java.util.List;
-
-@AutoService(InstrumentationModule.class)
-public class SpringPulsarInstrumentationModule extends InstrumentationModule {
-  public SpringPulsarInstrumentationModule() {
-    super("spring-pulsar", "spring-pulsar-1.2");
-  }
-
-  @Override
-  public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new DefaultPulsarMessageListenerContainerInstrumentation());
-  }
-}
