+++ settings.gradle.kts
@@ -552,8 +552,8 @@ include(":instrumentation:spring:spring-jms:spring-jms-6.0:javaagent")
 include(":instrumentation:spring:spring-kafka-2.7:javaagent")
 include(":instrumentation:spring:spring-kafka-2.7:library")
 include(":instrumentation:spring:spring-kafka-2.7:testing")
-include(":instrumentation:spring:spring-pulsar-1.2:javaagent")
-include(":instrumentation:spring:spring-pulsar-1.2:testing")
+include(":instrumentation:spring:spring-pulsar-1.0:javaagent")
+include(":instrumentation:spring:spring-pulsar-1.0:testing")
 include(":instrumentation:spring:spring-rabbit-1.0:javaagent")
 include(":instrumentation:spring:spring-rmi-4.0:javaagent")
 include(":instrumentation:spring:spring-scheduling-3.1:bootstrap")
